chorddiagOutput {chorddiag}	R Documentation

Widget output function for use in Shiny

Description

Widget output function for use in Shiny

Usage


chorddiagOutput(outputId, width = "90%", height = "350px")

Arguments

outputId
Shiny output variable to read from.

width
A valid CSS unit for the width of the chord diagram output.

height
A valid CSS unit for the height of the chord diagram output.

[Package chorddiag version 0.1.2 Index]